+++ Sources/Atomic.swift
@@ -155,7 +155,9 @@ internal class Lock {
 				attr.deallocate(capacity: 1)
 			}
 
-			#if DEBUG
+			// Darwin pthread for 32-bit ARM somehow returns `EAGAIN` when
+			// using `trylock` on a `PTHREAD_MUTEX_ERRORCHECK` mutex.
+			#if DEBUG && !arch(arm)
 			pthread_mutexattr_settype(attr, Int32(recursive ? PTHREAD_MUTEX_RECURSIVE : PTHREAD_MUTEX_ERRORCHECK))
 			#else
 			pthread_mutexattr_settype(attr, Int32(recursive ? PTHREAD_MUTEX_RECURSIVE : PTHREAD_MUTEX_NORMAL))